+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.3.x
 
 ## Bug fixes
+- PIM-3935: Fix "working copy value" tooltip is always empty
 - PIM-3957: Unable to publish a product already published
 
 # 1.3.5 (2015-03-19)